Issue #100: Integrate resize_inputs tests into preprocessing tests.
* Removed test_preprocess_resize.py
