+++ test/unit/test_preprocess.py
@@ -11,6 +11,7 @@ Are assumed working, and are tested separately in
 test_layer_util.py; as such we just check output size here.
 """
 import numpy as np
+import tensorflow as tf
 
 import deepreg.dataset.preprocess as preprocess
 
@@ -160,3 +161,53 @@ def test_transform_with_labels():
     assert outputs.get("moving_label").shape == moving_image_batched.shape
     assert outputs.get("fixed_label").shape == fixed_image_batched.shape
     assert len(outputs.get("indices")) == len(indices)
+
+
+def test_resize_inputs():
+    """
+   Test resize_inputs by confirming that it generates
+   appropriate output sizes on a simple test case.
+   """
+    input_size = (2, 2, 2)
+    moving_image_size = (1, 3, 5)
+    fixed_image_size = (2, 4, 6)
+
+    # labeled data - Pass
+    moving_image = tf.ones(input_size)
+    fixed_image = tf.ones(input_size)
+    moving_label = tf.ones(input_size)
+    fixed_label = tf.ones(input_size)
+    indices = tf.ones((2,))
+
+    inputs = dict(
+        moving_image=moving_image,
+        fixed_image=fixed_image,
+        moving_label=moving_label,
+        fixed_label=fixed_label,
+        indices=indices,
+    )
+    outputs = preprocess.resize_inputs(
+        inputs=inputs,
+        moving_image_size=moving_image_size,
+        fixed_image_size=fixed_image_size,
+    )
+
+    assert outputs["moving_image"].shape == moving_image_size
+    assert outputs["fixed_image"].shape == fixed_image_size
+    assert outputs["moving_label"].shape == moving_image_size
+    assert outputs["fixed_label"].shape == fixed_image_size
+
+    # unlabeled data - Pass
+    moving_image = tf.ones(input_size)
+    fixed_image = tf.ones(input_size)
+    indices = tf.ones((2,))
+
+    inputs = dict(moving_image=moving_image, fixed_image=fixed_image, indices=indices)
+    outputs = preprocess.resize_inputs(
+        inputs=inputs,
+        moving_image_size=moving_image_size,
+        fixed_image_size=fixed_image_size,
+    )
+
+    assert outputs["moving_image"].shape == moving_image_size
+    assert outputs["fixed_image"].shape == fixed_image_size
